開發商品列表

+++ ModelApplication/app/src/main/java/com/example/modelapplication/Product.java
@@ -6,8 +6,8 @@ public class Product {
     public String Description;
     public String Photo;
     public int CostJP  = 0;
-    public float CostExchangeRate  = 0f;
-    public float PriceExchangeRage  = 0f;
+    public double CostExchangeRate  = 0f;
+    public double PriceExchangeRage  = 0f;
     public int Inventory  = 0;
     public int OrderedQuantity  = 0;
     public String ProductTypeID;
